Use short int for rendering priority
It is quite unlikely that the game will have more than 32767 layers of
SceneObjects. short int also has -32768 to the negative side if there's
a need to go below zero.

+++ include/SceneObject.hpp
@@ -18,7 +18,7 @@ namespace Birb
 		SceneObject& operator=(SceneObject&&) 		= default;
 
 		/// The rendering priority determnies the order that objects are rendered with
-		int renderingPriority;
+		short renderingPriority;
 
 		/// Tells if the parent scene is active or not
 		bool sceneActive;
